refactor (LottoApplication) : step3 - LottoGame에 몰려있는 기능 분리 * Controller 기능 분리 * Service 기능 분리
