+++ src/main/java/lotto/LottoApplication.java
@@ -1,0 +1,9 @@
+package lotto;
+
+import lotto.controller.LottoController;
+
+public class LottoApplication {
+    public static void main(String[] args) {
+        LottoController.lottoGameStart();
+    }
+}
